move to depends; adopt dataclass, simplify assert
